Replace Gatsby with Vite
