+++ .github/workflows/build.yml
@@ -1,4 +1,4 @@
-name: Gatsby Test
+name: Build
 
 on: pull_request
 
@@ -7,5 +7,5 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v1
-      - run: yarn
-      - run: yarn build
+      - run: npm install
+      - run: npm run build
