+++ .nvmrc
@@ -1,1 +1,1 @@
-v14.15.0
+v18.17.0
