+++ package.json
@@ -1,57 +1,36 @@
 {
-  "name": "nulogy.design",
-  "version": "1.0.0",
+  "name": "content.nulogy.design",
   "private": true,
-  "description": "nulogy.design",
-  "author": "Matt Dunn",
-  "keywords": [
-    "gatsby"
-  ],
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "develop": "gatsby develop",
-    "precommit": "pretty-quick --staged",
-    "start": "rm -rf .cache/caches/gatsby-plugin-mdx && gatsby develop",
-    "build": "gatsby build",
-    "serve": "gatsby serve",
-    "clean": "gatsby clean",
-    "prettier": "prettier --write ."
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@mdx-js/mdx": "^1.6.22",
-    "@mdx-js/react": "^1.6.22",
-    "@nulogy/components": "^6.1.3",
-    "@nulogy/icons": "^4.17.0",
-    "assert": "^2.0.0",
-    "babel-plugin-styled-components": "^1.12.0",
-    "buffer": "^6.0.3",
-    "gatsby": "^3.0.0",
-    "gatsby-plugin-google-analytics": "^2.9.0",
-    "gatsby-plugin-manifest": "^3.3.0",
-    "gatsby-plugin-mdx": "^1.8.0",
-    "gatsby-plugin-prettier-eslint": "^1.0.4",
-    "gatsby-plugin-react-svg": "^3.0.0",
-    "gatsby-plugin-sharp": "^3.4.0",
-    "gatsby-plugin-styled-components": "^4.4.0",
-    "gatsby-plugin-webfonts": "^2.1.0",
-    "gatsby-remark-images": "^5.1.0",
-    "gatsby-source-filesystem": "^2.9.1",
-    "os-browserify": "^0.3.0",
-    "path-browserify": "^1.0.1",
-    "prism-react-renderer": "^1.1.1",
-    "raw-loader": "^4.0.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-helmet": "^6.1.0",
-    "react-view": "^2.3.7",
-    "styled-components": "^5.2.1"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "yarn run precommit"
-    }
+    "@nulogy/components": "^8.18.0",
+    "@nulogy/icons": "^4.26.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.20.1",
+    "styled-components": "^5.3.11",
+    "stylis": "^4.3.0"
   },
   "devDependencies": {
-    "husky": "^6.0.0",
-    "prettier": "2.3.2"
+    "@mdx-js/react": "^3.0.0",
+    "@mdx-js/rollup": "^3.0.0",
+    "@types/react": "^18.2.37",
+    "@types/react-dom": "^18.2.15",
+    "@types/styled-components": "^5.1.34",
+    "@typescript-eslint/eslint-plugin": "^6.10.0",
+    "@typescript-eslint/parser": "^6.10.0",
+    "@vitejs/plugin-react": "^4.2.0",
+    "eslint": "^8.53.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.4",
+    "typescript": "^5.2.2",
+    "vite": "^5.0.0"
   }
 }
