+++ src/App.tsx
@@ -1,0 +1,6 @@
+import { RouterProvider } from "react-router-dom";
+import router from "./Router"
+
+export default function App() {
+	return <RouterProvider router={router} />;
+}
