--- src/components/CONSTANTS.js
@@ -1,9 +1,0 @@
-export const CODE_WIDTH = "1024px";
-export const COPY_WIDTH = "640px";
-
-export const FIGMA_LINK =
-  "https://www.figma.com/file/IH4gk9P2E4xEYiNKoD22Jd/NDS-v1?node-id=105%3A48";
-export const GITHUB_LINK = "https://github.com/nulogy/design-system";
-export const SLACK_LINK = "slack://channel?team=T024N2KKA&id=CBAFQ4X7X";
-export const STORYBOOK_LINK =
-  "https://master--5f60c6c285eaad0022dce67f.chromatic.com/";
