+++ src/components/CONSTANTS.ts
@@ -1,0 +1,2 @@
+export const CODE_WIDTH = "1024px";
+export const COPY_WIDTH = "640px";
