--- src/data/routes.js
@@ -1,90 +1,0 @@
-/* This Array configures the the sections and name, href, and title for each page on the website.
-For the page title that appears in the browser tab, the name will be used unless a title is specified. */
-
-export const routes = [
-  {
-    name: "Voice and tone",
-    links: [
-      {
-        name: "Overview",
-        title: "Voice and tone",
-        href: "/voice-and-tone/overview",
-      },
-      {
-        name: "Voice",
-        href: "/voice-and-tone/voice",
-      },
-      {
-        name: "Tone",
-        href: "/voice-and-tone/tone",
-      },
-    ],
-  },
-  {
-    name: "Grammar and mechanics",
-    links: [
-      {
-        name: "Tenses",
-        href: "/grammar-and-mechanics/tenses",
-      },
-      {
-        name: "Capitalization",
-        href: "/grammar-and-mechanics/capitalization",
-      },
-      {
-        name: "Headings",
-        href: "/grammar-and-mechanics/headings",
-      },
-      {
-        name: "Lists",
-        href: "/grammar-and-mechanics/lists",
-      },
-      {
-        name: "Numbers",
-        href: "/grammar-and-mechanics/numbers",
-      },
-      {
-        name: "Contractions and abbreviations",
-        href: "/grammar-and-mechanics/contractions-and-abbreviations",
-      },
-      {
-        name: "Symbols",
-        href: "/grammar-and-mechanics/symbols",
-      },
-      {
-        name: "Dates and times",
-        href: "/grammar-and-mechanics/dates-and-times",
-      },
-      {
-        name: "Web elements",
-        href: "/grammar-and-mechanics/web-elements",
-      },
-    ],
-  },
-  {
-    name: "Writing for translation",
-    links: [
-      {
-        name: "Overview",
-        title: "Writing for translation",
-        href: "/writing-for-translation/overview",
-      },
-      {
-        name: "Be concise",
-        href: "/writing-for-translation/be-concise",
-      },
-      {
-        name: "Be clear",
-        href: "/writing-for-translation/be-clear",
-      },
-      {
-        name: "Use active voice",
-        href: "/writing-for-translation/use-active-voice",
-      },
-      {
-        name: "Use positive voice",
-        href: "/writing-for-translation/use-positive-voice",
-      },
-    ],
-  },
-];
