+++ src/pages/404.tsx
@@ -1,0 +1,12 @@
+import { Box } from "@nulogy/components";
+
+const PageNotFound = () => {
+  return (
+    <Box p="x4">
+      <h1>Page not found</h1>
+      <p>The page you're looking for is not here.</p>
+    </Box>
+  );
+};
+
+export default PageNotFound;
