+++ src/pages/grammar-and-mechanics/capitalization.mdx
@@ -8,6 +8,7 @@ Some general rules:
 
 Headings and subheadings use sentence case.
 
+
 <NoExample>Edit Production Records</NoExample>
 <YesExample>Edit production records</YesExample>
 
@@ -15,12 +16,8 @@ Headings and subheadings use sentence case.
 
 Use title case for the names of our products
 
-<NoExample>
-  The ecosystem solution allows brands insights into the production process.
-</NoExample>
-<YesExample>
-  The Ecosystem Solution allows brands insights into the production process.
-</YesExample>
+<NoExample>The ecosystem solution allows brands insights into the production process.</NoExample>
+<YesExample>The Ecosystem Solution allows brands insights into the production process.</YesExample>
 
 ## Tools and functions
