+++ src/pages/grammar-and-mechanics/contractions-and-abbreviations.mdx
@@ -1,5 +1,5 @@
 import { Flex, Box, Text } from "@nulogy/components";
-import { COPY_WIDTH } from "../../components/CONSTANTS.js";
+import { COPY_WIDTH } from "../../components/CONSTANTS.ts";
 
 # Contractions and abbreviations
