+++ src/pages/grammar-and-mechanics/dates-and-times.mdx
@@ -1,5 +1,5 @@
 import { Flex, Box, Text } from "@nulogy/components";
-import { COPY_WIDTH } from "../../components/CONSTANTS.js";
+import { COPY_WIDTH } from "../../components/CONSTANTS.ts";
 
 # Dates and times
 
@@ -57,7 +57,7 @@ Given that the ISO date format is not fully adopted worldwide, Nulogy goes a ste
   </Box>
 </Flex>
 
-For example, November 4th, 2021 would read as 2021-Nov-04. July 28th, 2025 would read as 2025-Jul-28. 
+For example, November 4th, 2021 would read as 2021-Nov-04. July 28th, 2025 would read as 2025-Jul-28.
 
 <Flex justifyContent="space-between" maxWidth={COPY_WIDTH} m="0 auto" mb="x4">
   <Box>
@@ -155,7 +155,7 @@ When a date and time are listed together, we list the date first, followed by th
   </Box>
 </Flex>
 
-In a full sentence, we use additional words such as “on” and “at” to help readability. 
+In a full sentence, we use additional words such as “on” and “at” to help readability.
 
 <NoExample>
   This bill of materials was edited by Jenna Newton at 19:22 Nov 4, 2021.
@@ -165,7 +165,7 @@ In a full sentence, we use additional words such as “on” and “at” to hel
 </YesExample>
 
 ## Ranges
-When listing a date or time range, the beginning date/time and ending date/time are seperated by the word “to”. This is to avoid confusion with the dashes from the standalone date formating. 
+When listing a date or time range, the beginning date/time and ending date/time are seperated by the word “to”. This is to avoid confusion with the dashes from the standalone date formating.
 
 Standalone:
 
@@ -205,7 +205,7 @@ In a sentence:
 </YesExample>
 
 <NoExample>
-  Nulogy’s vision retreat will be from 9:00 AM - 10:00 PM on January 20th, 2022. 
+  Nulogy’s vision retreat will be from 9:00 AM - 10:00 PM on January 20th, 2022.
 </NoExample>
 <YesExample>
   Nulogy’s vision retreat will be held on January 20th, 2022 from 9:00 AM to 10:00 PM.
