+++ src/pages/grammar-and-mechanics/headings.mdx
@@ -1,3 +1,5 @@
+import { Heading2 } from "@nulogy/components";
+
 # Headings and subheadings
 
 Headings should always be used to describe any text or interactions that can be achieved in the body below. They should be short and descriptive.
