+++ src/pages/grammar-and-mechanics/lists.mdx
@@ -1,5 +1,5 @@
 import { Box, Flex, List, ListItem, Text } from "@nulogy/components";
-import { COPY_WIDTH } from "../../components/CONSTANTS.js";
+import { COPY_WIDTH } from "../../components/CONSTANTS.ts";
 
 # Lists
 
@@ -18,11 +18,8 @@ Lists can be either ordered or unordered. Ordered lists (numbered points 1, 2, 3
   </Box>
   <Box>
     <YesExample>
-      <Text fontSize="small" mt="x2">
-        To assign pallets through item locator:
-      </Text>
-      <List fontSize="small" listStyle="order">
-        <ListItem>To assign pallets through item locator:</ListItem>
+      <Text fontSize="small" mt="x2">To assign pallets through item locator: </Text>
+      <List marginTop="x2" fontSize="small" listStyle="order">
         <ListItem>Choose a work order.</ListItem>
         <ListItem>Review pallets to be assigned.</ListItem>
         <ListItem>Click the "assign" button</ListItem>
@@ -36,9 +33,7 @@ If a list item is a full sentence, then it should include proper punctuation. Li
 <Flex justifyContent="space-between" maxWidth={COPY_WIDTH} m="0 auto" mb="x4">
   <Box>
     <NoExample>
-      <Text fontSize="small" mt="x2">
-        Attachment information includes:
-      </Text>
+      <Text fontSize="small" mt="x2">Attachment information includes:</Text>
       <List fontSize="small" mt="x2">
         <ListItem>File name;</ListItem>
         <ListItem>Upload date;</ListItem>
@@ -48,9 +43,7 @@ If a list item is a full sentence, then it should include proper punctuation. Li
   </Box>
   <Box>
     <YesExample>
-      <Text fontSize="small" mt="x2">
-        Attachment information includes:
-      </Text>
+      <Text fontSize="small" mt="x2">Attachment information includes:</Text>
       <List fontSize="small" mt="x2">
         <ListItem>File name</ListItem>
         <ListItem>Upload date</ListItem>
@@ -62,21 +55,15 @@ If a list item is a full sentence, then it should include proper punctuation. Li
 
 When three or more items are listed in a sentence we use the Oxford comma, which means that we include a comma before the “and” (e.g. This content guide includes sections on voice, tone, and grammar).
 
-<NoExample fontSize="small">
-  This content guide includes section on voice, tone and grammar.
-</NoExample>
-<YesExample fontSize="small" mb="x4">
-  This content guide includes sections on voice, tone, and grammar.
-</YesExample>
+<NoExample fontSize="small">This content guide includes section on voice, tone and grammar.</NoExample>
+<YesExample fontSize="small" mb="x4">This content guide includes sections on voice, tone, and grammar.</YesExample>
 
 Lists can vary in format, but all items within the same list should match format. For example, all sentence fragments, all beginning with a verb, all the same length.
 
 <Flex justifyContent="space-between" maxWidth={COPY_WIDTH} m="0 auto" mb="x4">
   <Box>
     <NoExample>
-      <Text fontSize="small" mt="x2">
-        Use Nulogy's Shop Floor Control to report on:
-      </Text>
+      <Text fontSize="small" mt="x2">Use Nulogy's Shop Floor Control to report on:</Text>
       <List fontSize="small" mt="x2">
         <ListItem>
           Inventory adjustments made by individual staff members.
@@ -88,9 +75,7 @@ Lists can vary in format, but all items within the same list should match format
   </Box>
   <Box>
     <YesExample>
-      <Text fontSize="small" mt="x2">
-        Use Nulogy's Shop Floor Controp to report on:
-      </Text>
+      <Text fontSize="small" mt="x2">Use Nulogy's Shop Floor Controp to report on:</Text>
       <List fontSize="small" mt="x2">
         <ListItem>Inventory adjustments</ListItem>
         <ListItem>Item reservations</ListItem>
