+++ src/pages/grammar-and-mechanics/numbers.mdx
@@ -1,10 +1,12 @@
+import { Text } from "@nulogy/components"
+
 # Numbers
 
 Use numeric digits if the number in question is exact. The shape of the number helps readers to identify it. Numeric digits also don’t require translation. If it is a generalization, spell it out.
 
 <YesExample>
-  Your search generated <strong>18</strong> results.
+    <Text>Your search generated <strong>18</strong> results.</Text>
 </YesExample>
 <YesExample>
-  <strong>Thousands</strong> of companies use Nulogy.
+    <Text><strong>Thousands</strong> of companies use Nulogy.</Text>
 </YesExample>
