+++ src/pages/grammar-and-mechanics/tenses.mdx
@@ -2,10 +2,6 @@
 
 Always use the present tense unless the referred action has specifically taken place in the past or will take place in the future. This makes our voice immediate, which works best because our customers work in an immediate environment.
 
-<YesExample>
-  There <strong>were</strong> 8 errors on this form
-</YesExample>
+<YesExample>There <strong>were</strong> 8 errors on this form</YesExample>
 
-<NoExample>
-  There <strong>are</strong> 8 errors on this form
-</NoExample>
+<NoExample>There <strong>are</strong> 8 errors on this form</NoExample>
