+++ src/pages/grammar-and-mechanics/web-elements.mdx
@@ -1,5 +1,5 @@
 import { Box, Checkbox, Flex, PrimaryButton } from "@nulogy/components";
-import { COPY_WIDTH } from "../../components/CONSTANTS.js";
+import { COPY_WIDTH } from "../../components/CONSTANTS.ts";
 
 # Web elements
