+++ src/pages/voice-and-tone/voice.mdx
@@ -1,9 +1,9 @@
 import { Text } from "@nulogy/components";
-import { COPY_WIDTH } from "../../components/CONSTANTS.js";
+import { COPY_WIDTH } from "../../components/CONSTANTS.ts";
 
 # Voice
 
-Our voice is derived from our company values and from our North Star:
+The voice is derived from our company values and from our North Star:
 
 <Text
   maxWidth={COPY_WIDTH}
