+++ src/pages/writing-for-translation/be-clear.mdx
@@ -1,7 +1,3 @@
----
-title: Be clear
----
-
 # Be clear
 
 While we want to keep our language concise, certain words that in English appear as if they can be cut will help with clarity if they are included.
