+++ src/pages/writing-for-translation/be-concise.mdx
@@ -1,6 +1,3 @@
----
-title: Be concise
----
 
 # Be concise
