+++ src/pages/writing-for-translation/overview.mdx
@@ -1,6 +1,4 @@
----
-title: Overview
----
+
 
 # Writing for translation
