+++ src/pages/writing-for-translation/use-active-voice.mdx
@@ -1,6 +1,3 @@
----
-title: Use the active voice
----
 
 # Use the active voice
