+++ src/pages/writing-for-translation/use-positive-voice.mdx
@@ -1,7 +1,3 @@
----
-title: Use positive words
----
-
 # Use positive words
 
 Positive sentences are good for more than just the encouragement of our readers and users. The English language gives way to easily use negative words to communicate positive thought. Double negatives, or sentences that begin with phrases like “don’t you think”, or “can’t you see” infuse negative terms into otherwise positive sentences. When translated, this does not read in the same way. It is, therefore, best practice to use the positive voice wherever possible.
